[#13] refactor: #12 PR 피드백 반영
